Changes according to feedback
