+++ src/kubectl.ts
@@ -5,7 +5,9 @@ import { FS } from './fs';
 import { Shell, ShellHandler, ShellResult } from './shell';
 import * as binutil from './binutil';
 import { Errorable } from './errorable';
-import { parseLineOutput, POD_OUTPUT_PARSING_REGEX } from './kubectlUtils';
+import { parseLineOutput } from './outputUtils';
+
+const KUBECTL_OUTPUT_COLUMN_SEPARATOR = /\s+/g;
 
 export interface Kubectl {
     checkPresent(errorMessageMode: CheckPresentMessageMode): Promise<boolean>;
@@ -237,7 +239,7 @@ async function fromLines(context: Context, command: string): Promise<Errorable<{
     if (shellResult.code === 0) {
         let lines = shellResult.stdout.split('\n');
         lines = lines.filter((l) => l.length > 0);
-        const parsedOutput = parseLineOutput(lines, POD_OUTPUT_PARSING_REGEX);
+        const parsedOutput = parseLineOutput(lines, KUBECTL_OUTPUT_COLUMN_SEPARATOR);
         return { succeeded: true, result: parsedOutput };
     }
     return { succeeded: false, error: [ shellResult.stderr ] };
