+++ src/kubectlUtils.ts
@@ -5,8 +5,6 @@ import { sleep } from "./sleep";
 import { ObjectMeta, KubernetesCollection, DataResource, Namespace, Pod, KubernetesResource } from './kuberesources.objectmodel';
 import { failed } from "./errorable";
 
-export const POD_OUTPUT_PARSING_REGEX = /\s+/g;
-
 export interface KubectlContext {
     readonly clusterName: string;
     readonly contextName: string;
@@ -335,27 +333,4 @@ export async function getResourceAsJson<T extends KubernetesResource | Kubernete
         return undefined;
     }
     return shellResult.result;
-}
-
-/**
- * Parse column based output which is seperated by whitespace(s) from kubectl or similar sources
- * for example, kubectl get po
- * @param lineOutput raw output with headers from kubectl or similar sources
- * @param parsingRegex regex to parse the columns
- * @return array of objects with key as column header and value
- */
-export function parseLineOutput(lineOutput: string[], parsingRegex: RegExp): { [key: string]: string }[] {
-    const headers = lineOutput.shift();
-    if (!headers) {
-        return [];
-    }
-    const parsedHeaders = headers.toLowerCase().replace(parsingRegex, '|').split('|');
-    return lineOutput.map((line) => {
-        const lineInfoObject = {};
-        const bits = line.replace(parsingRegex, '|').split('|');
-        bits.forEach((columnValue, index) => {
-            lineInfoObject[parsedHeaders[index].trim()] = columnValue.trim();
-        });
-        return lineInfoObject;
-    });
 }
